refactor: unite read from socket with RMS processing - move RpcCommand structure to tcp.rs - change reading to single function and impl for RpcCommand - do processing in single task
